+++ src/frontend/web_application/src/modules/a11y/services/tabIndexes.js
@@ -1,0 +1,1 @@
+export const draftExcerptIndex = 0;